refactor: move mutations into own directory
